Refactored zoom and coordinate systems
Project-wide refactoring: Instead of one coordinate system that is 
updated every time you zoom, there are now two coordinate systems m and 
v, m (for model) doesn't change when zooming, and v (for view)
does change. All functions and variables dealing with the two are 
renamed to reflect whether they apply to m or v coordinates.

There are still minor tweaks to be made, but other than that this is a 
complete refactoring.